REBASEME: import Xiaomi camera changes

+++ drivers/media/platform/msm/camera/cam_cdm/cam_cdm.h
@@ -1,4 +1,5 @@
 /* Copyright (c) 2017-2018, The Linux Foundation. All rights reserved.
+ * Copyright (C) 2020 XiaoMi, Inc.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 and
@@ -199,7 +200,7 @@ struct cam_cdm_hw_intf_cmd_submit_bl {
 struct cam_cdm_hw_mem {
 	int32_t handle;
 	uint32_t vaddr;
-	uintptr_t kmdvaddr;
+	uint64_t kmdvaddr;
 	size_t size;
 };
